Upgraded to 256 MB and fixed signed integer parsing

+++ src/profile/profile.cpp
@@ -42,12 +42,12 @@ Profile loadProfile(std::string profile_path) {
 
     if (config.contains("starting-pc")) {
         std::string sp = config["starting-pc"];
-        profile.starting_pc = std::stoi(sp, nullptr, 16);
+        profile.starting_pc = static_cast<uint32_t>(std::stoul(sp, nullptr, 16));
     }
 
     if (config.contains("dtb-address")) {
         std::string da = config["dtb-address"];
-        profile.dtb_address = std::stoi(da, nullptr, 16);
+        profile.dtb_address = static_cast<uint32_t>(std::stoul(da, nullptr, 16));
     }
 
     if (config.contains("extensions") && config["extensions"].is_array()) {
